+++ apis/al.py
@@ -1,5 +1,4 @@
 from fastapi import APIRouter, HTTPException
-from models.ticket import Ticket,TicketDB
 from modules.database import SessionLocal
 import lorem
 from lorem.text import TextLorem
@@ -9,7 +8,6 @@ router = APIRouter()
 
 from uuid import uuid4
 import random
-
 
 from modules import utils
 
@@ -52,12 +50,10 @@ async def generate_quote(mode: str='paragraph'):
             
     return utils.response({"quote":text})
 
-
 @router.get('/random_first_name')
 async def randomFirstName():
     return utils.response({"first_name":names.get_first_name()})
     
-
 @router.get('/random_last_name')
 async def randomLastName():
     return utils.response({"last_name":names.get_last_name()})
@@ -65,20 +61,3 @@ async def randomLastName():
 @router.get('/root_directory')
 async def get_root():
     return utils.response({"directory":utils.get_project_root()})
-
-@router.post('/submit_ticket')
-async def postSubmitTicket(ticket:Ticket):
-    db = SessionLocal()
-    ticket_item = TicketDB(**ticket.dict())
-    db.add(ticket_item)
-    
-    
-    try:
-        db.commit()
-    except Exception as e:
-        db.rollback()
-        raise HTTPException(status_code=500,detail=str(e))
-    finally:
-        db.close()
-    
-    return utils.response({"message":"Ticket submitted correctly aye","ticket_item":ticket})
